refactor: remove final deprecated functions, fields, and re-export shims
- Delete deprecated ErrorResponse() from pkg/i18n (replaced by pkg/apierr)
- Delete deprecated setupWebhook() from repository API (duplicate of registerWebhook)
- Remove deprecated SetCallback() from terminal state detector interface
  and all implementations; Subscribe() is the only notification method
- Remove deprecated PodCreateRequest fields (Prompt, Model, ConfigOverrides,
  PermissionMode); all pod config flows through AgentFile exclusively
- Remove 7 backward-compat re-export shims from web frontend:
  useChannelChat, ticket API types, channelStore, CreatePodModal,
  WebhookSettings, ImportRepositoryModal, AgentPanel
