+++ web/src/app/(dashboard)/[org]/repositories/[id]/components/index.ts
@@ -8,7 +8,6 @@ export { useRepositoryDetail, type RepositoryTab } from "./useRepositoryDetail";
 export { RepositoryHeader } from "./RepositoryHeader";
 export { RepositoryInfoCard } from "./RepositoryInfoCard";
 export { GitProviderCard } from "./GitProviderCard";
-export { RepositoryActionsCard } from "./RepositoryActionsCard";
 export { RepositoryTabs } from "./RepositoryTabs";
 export { WebhookSettingsCard } from "./WebhookSettingsCard";
 export { CapabilitiesTab } from "./CapabilitiesTab";
