--- web/src/components/repository/WebhookSettings.tsx
@@ -1,3 +1,0 @@
-// Re-export from webhook folder for backward compatibility
-export { WebhookSettings, type WebhookSettingsProps } from "./webhook";
-export { WebhookSettings as default } from "./webhook";
